First commit: Basic CPP expression parser
A very basic implementation of a CPP expression parser.  This should be
jammed into a more general CPP parser which handles the actual
statements (#include, ``#ifdef, etc).

Implementation follows a shunting yard algorithm.  This is not the usual
Dijkstra method which uses a second operator stack and then converts the
whole thing into postfix format.  This is the "other" method that
immediately applies any preceding operators  which are of equal weight
to the current operator.

(I'm not a classically trained programmer, so I don't know what you call
this approach.)

Far from perfect, far from efficient, and far from complete.  Surely has
its holes and broken cases.  But if it solves our makedep issues then
it's good enough.